# Conflicts:
#	server/src/main/java/com/server/app/enums/ResultMessages.java
#	server/src/main/java/com/server/app/service/EmployeeService.java
#	server/src/main/java/com/server/app/service/ShipperService.java
#	server/src/main/java/com/server/app/service/srvImpl/SupplierSrvImpl.java

+++ server/src/main/java/com/server/app/dto/TerritoryDto.java
@@ -1,6 +1,5 @@
 package com.server.app.dto;
 
-import com.server.app.model.Region;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class TerritoryDto {
     private String territoryId;
     private String territoryDescription;
-    private Short regionId;
+    private Long regionId;
 }
