+++ server/src/main/java/com/server/app/dto/request/TerritoryUpdateRequest.java
@@ -3,7 +3,6 @@ package com.server.app.dto.request;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import com.server.app.model.Region;
 
 @Data
 @NoArgsConstructor
@@ -11,5 +10,5 @@ import com.server.app.model.Region;
 public class TerritoryUpdateRequest {
     private String territoryId;
     private String territoryDescription;
-    private Short regionId;
+    private Long regionId;
 }
